requirements.txt
absl-py==0.7.0
aiohttp==3.5.4
altgraph==0.16.1
apiai==1.2.3
arcade==1.3.7
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
attrs==18.2.0
audioread==2.1.6
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.2
cycler==0.10.0
DateTime==4.3
DAWG-Python==0.7.2
decorator==4.3.2
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.3
Flask==1.0.2
future==0.17.1
gast==0.2.2
get==2019.4.13
giphy-client==1.0.0
giphypop==0.3
googletrans==2.4.0
graphviz==0.10.1
grpcio==1.18.0
h5py==2.9.0
html2text==2018.1.9
html5lib==1.0.1
idna==2.8
IMDbPY==6.8
instaloader==4.2.3
instaparser==1.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
joblib==0.13.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kaggle==1.5.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.27.0
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mechanize==0.4.0
mistune==0.8.4
mock==2.0.0
mpmath==1.1.0
multidict==4.5.2
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.4
notebook==5.7.4
numba==0.42.0
numpy==1.16.0
pa==0.0.23.dev0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.2
pbr==5.1.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==5.3.0
pipenv==2018.11.26
playsound==1.2.2
post==2019.4.13
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.1
public==2019.4.13
PyAudio==0.2.11
pycparser==2.19
pydnevnikruapi==0.0.7
pydot==1.4.1
pydotplus==2.0.2
pydub==0.23.1
pygame==1.9.4
pyglet==1.3.2
Pygments==2.3.1
PyInstaller==3.5
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyparsing==2.3.1
pypiwin32==223
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyqt5-tools==5.11.2.1.3
PySocks==1.6.8
pyTelegramBotAPI==3.6.6
python-box==3.4.0
python-dateutil==2.8.0
python-dotenv==0.9.1
python-slugify==2.0.1
python-telegram-bot==11.1.0
pytz==2018.9
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
qrcode==6.0
qreader==0.1
qtconsole==4.4.3
query-string==2019.4.13
reedsolo==0.3
request==2019.4.13
requests==2.21.0
resampy==0.2.1
rutermextract==0.3
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.7.2
SQLAlchemy==1.3.6
sympy==1.3
task==0.0.0
telepot==12.7
tensorboard==1.12.2
tensorflow==1.13.0rc1
tensorflow-estimator==1.13.0rc0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tmdb3==0.7.2
tmdbsimple==2.2.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.22
virtualenv==16.2.0
virtualenv-clone==0.4.0
voluptuous==0.11.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
yandex-weather-api==0.2.4
yarl==1.3.0
zope.interface==4.6.0
